configmgr: add SAL_CONFIG_WINREG_RETAIN_TMP variable.
Windows registry dumping behaves very strangely, and in hard to debug
ways - this should help in-the-field debugging of this in future.

Reviewed-on: https://gerrit.libreoffice.org/33992
Tested-by: Michael Meeks <michael.meeks@collabora.com>
(cherry picked from commit e5b062758950b0d6c5cea0a1914998433ef19774)

+++ configmgr/source/components.cxx
@@ -561,7 +561,8 @@ Components::Components(
             OUString aTempFileURL;
             if (dumpWindowsRegistry(&aTempFileURL, eType)) {
                 parseFileLeniently(&parseXcuFile, aTempFileURL, layer, 0, 0, 0);
-                osl::File::remove(aTempFileURL);
+                if (!getenv("SAL_CONFIG_WINREG_RETAIN_TMP"))
+                    osl::File::remove(aTempFileURL);
             }
             ++layer; //TODO: overflow
 #endif
